Move GitHub API in its own namespace
