+++ lib/cucumber/website/git_hub/fake.rb
@@ -1,0 +1,11 @@
+module Cucumber
+  module Website
+    module GitHub
+
+      class Fake
+        attr_accessor :events
+      end
+    end
+
+  end
+end
